Move long lists of constants to separate files
