+++ constants.php
@@ -1,4 +1,8 @@
 <?php 
+require('constants/math.php');
+require('constants/mistakes.php');
+require('constants/parameters.php');
+
 define('HOME', dirname(__FILE__) . '/');
 
 const PIPE_PLACEHOLDER = '%%CITATION_BOT_PIPE_PLACEHOLDER%%';
@@ -21,72 +25,9 @@ const HTML_ENCODE = array("&#x5B;", "&#x5D;", "&#60;", "&#62;", "+");
 const DOT_ENCODE = array(".2F", ".5B", ".7B", ".7D", ".5D", ".3C", ".3E", ".3B", ".28", ".29");
 const DOT_DECODE = array("/", "[", "{", "}", "]", "<", ">", ";", "(", ")");
 
-const COMMON_MISTAKES = array ( // Common mistakes that aren't picked up by the levenshtein approach
-  "albumlink"       =>  "titlelink",
-  "artist"          =>  "others",
-  "authorurl"       =>  "authorlink",
-  "co-author"       =>  "coauthor",
-  "co-authors"      =>  "coauthors",
-  "dio"             =>  "doi",
-  "director"        =>  "others",
-  "display-authors" =>  "displayauthors",
-  "display_authors" =>  "displayauthors",
-  "doi_brokendate"  =>  "doi-broken-date",
-  "doi_inactivedate"=>  "doi-broken-date",
-  "doi-inactive-date"   =>  "doi-broken-date",
-  "ed"              =>  "editor",
-  "ed2"             =>  "editor2",
-  "ed3"             =>  "editor3",
-  "editorlink1"     =>  "editor1-link",
-  "editorlink2"     =>  "editor2-link",
-  "editorlink3"     =>  "editor3-link",
-  "editorlink4"     =>  "editor4-link",
-  "editor1link"     =>  "editor1-link",
-  "editor2link"     =>  "editor2-link",
-  "editor3link"     =>  "editor3-link",
-  "editor4link"     =>  "editor4-link",
-  "editor-first1"   =>  "editor1-first",
-  "editor-first2"   =>  "editor2-first",
-  "editor-first3"   =>  "editor3-first",
-  "editor-first4"   =>  "editor4-first",
-  "editor-last1"    =>  "editor1-last",
-  "editor-last2"    =>  "editor2-last",
-  "editor-last3"    =>  "editor3-last",
-  "editor-last4"    =>  "editor4-last",
-  "editorn"         =>  "editor2",
-  "editorn-link"    =>  "editor2-link",
-  "editorn-last"    =>  "editor2-last",
-  "editorn-first"   =>  "editor2-first",
-  "firstn"          =>  "first2",
-  "ibsn"            =>  "isbn",
-  "ibsn2"           =>  "isbn",
-  "lastn"           =>  "last2",
-  "part"            =>  "issue",
-  "no"              =>  "issue",
-  "No"              =>  "issue",
-  "No."             =>  "issue",
-  "notestitle"      =>  "chapter",
-  "nurl"            =>  "url",
-  "origmonth"       =>  "month",
-  "p"               =>  "page",
-  "p."              =>  "page",
-  "pmpmid"          =>  "pmid",
-  "pp"              =>  "pages",
-  "pp."             =>  "pages",
-  "publisherid"     =>  "id",
-  "titleyear"       =>  "origyear",
-  "translator"      =>  "others",
-  "translators"     =>  "others",
-  "vol"             =>  "volume",
-  "Vol"             =>  "volume",
-  "Vol."            =>  "volume",
-  "website"         =>  "url",
-);
-
 
 // dontCap is am array of strings that should not be capitalized in their titlecase format; 
 // unCapped is their correct capitalization. Remember to enclose any word in spaces.
-
 
 const LC_SMALL_WORDS = array(' and then ', ' of ',' the ',' and ',' an ',' or ',' nor ',' but ',' is ',' if ',' then ',' else ',' when', 'at ',' from ',' by ',' on ',' off ',' for ',' in ',' over ',' to ',' into ',' with ',' U S A ',' USA ',' et ');
 const UC_SMALL_WORDS = array(' and Then ', ' Of ',' The ',' And ',' An ',' Or ',' Nor ',' But ',' Is ',' If ',' Then ',' Else ',' When', 'At ',' From ',' By ',' On ',' Off ',' For ',' In ',' Over ',' To ',' Into ',' With ',' U S A ',' Usa ',' Et ');
@@ -107,785 +48,5 @@ const UCFIRST_JOURNAL_ACRONYMS = array(
 ' Molecular And Cellular Biology ', ' Ocean Science Journal : Osj ', ' Palaios ', ' Plos One ', ' Pnas ',
 ' S.a.p.i.en.s ', ' Star Trek: The Official Monthly Magazine ', ' The Embo Journal ', ' Time Out London ',
 ' Z/journal ', ' Zeitschrift Für Physik A Hadrons And Nuclei ', ' Zeitschrift Für Physik A: Hadrons And Nuclei ');
-
-const AUTHOR_PARAMETERS = array(
-    1  => array('surname'  , 'forename'  , 'initials'  , 'first'  , 'last'  , 'author',
-                'surname1' , 'forename1' , 'initials1' , 'first1' , 'last1' , 'author1', 'authors', 'vauthors'),
-    2  => array('surname2' , 'forename2' , 'initials2' , 'first2' , 'last2' , 'author2' , 'coauthors', 'coauthor'),
-    3  => array('surname3' , 'forename3' , 'initials3' , 'first3' , 'last3' , 'author3' ),
-    4  => array('surname4' , 'forename4' , 'initials4' , 'first4' , 'last4' , 'author4' ),
-    5  => array('surname5' , 'forename5' , 'initials5' , 'first5' , 'last5' , 'author5' ),
-    6  => array('surname6' , 'forename6' , 'initials6' , 'first6' , 'last6' , 'author6' ),
-    7  => array('surname7' , 'forename7' , 'initials7' , 'first7' , 'last7' , 'author7' ),
-    8  => array('surname8' , 'forename8' , 'initials8' , 'first8' , 'last8' , 'author8' ),
-    9  => array('surname9' , 'forename9' , 'initials9' , 'first9' , 'last9' , 'author9' ),
-    10 => array('surname10', 'forename10', 'initials10', 'first10', 'last10', 'author10'),
-    11 => array('surname11', 'forename11', 'initials11', 'first11', 'last11', 'author11'),
-    12 => array('surname12', 'forename12', 'initials12', 'first12', 'last12', 'author12'),
-    13 => array('surname13', 'forename13', 'initials13', 'first13', 'last13', 'author13'),
-    14 => array('surname14', 'forename14', 'initials14', 'first14', 'last14', 'author14'),
-    15 => array('surname15', 'forename15', 'initials15', 'first15', 'last15', 'author15'),
-    16 => array('surname16', 'forename16', 'initials16', 'first16', 'last16', 'author16'),
-    17 => array('surname17', 'forename17', 'initials17', 'first17', 'last17', 'author17'),
-    18 => array('surname18', 'forename18', 'initials18', 'first18', 'last18', 'author18'),
-    19 => array('surname19', 'forename19', 'initials19', 'first19', 'last19', 'author19'),
-    20 => array('surname20', 'forename20', 'initials20', 'first20', 'last20', 'author20'),
-    21 => array('surname21', 'forename21', 'initials21', 'first21', 'last21', 'author21'),
-    22 => array('surname22', 'forename22', 'initials22', 'first22', 'last22', 'author22'),
-    23 => array('surname23', 'forename23', 'initials23', 'first23', 'last23', 'author23'),
-    24 => array('surname24', 'forename24', 'initials24', 'first24', 'last24', 'author24'),
-    25 => array('surname25', 'forename25', 'initials25', 'first25', 'last25', 'author25'),
-    26 => array('surname26', 'forename26', 'initials26', 'first26', 'last26', 'author26'),
-    27 => array('surname27', 'forename27', 'initials27', 'first27', 'last27', 'author27'),
-    28 => array('surname28', 'forename28', 'initials28', 'first28', 'last28', 'author28'),
-    29 => array('surname29', 'forename29', 'initials29', 'first29', 'last29', 'author29'),
-    30 => array('surname30', 'forename30', 'initials30', 'first30', 'last30', 'author30'),
-    31 => array('surname31', 'forename31', 'initials31', 'first31', 'last31', 'author31'),
-    32 => array('surname32', 'forename32', 'initials32', 'first32', 'last32', 'author32'),
-    33 => array('surname33', 'forename33', 'initials33', 'first33', 'last33', 'author33'),
-    34 => array('surname34', 'forename34', 'initials34', 'first34', 'last34', 'author34'),
-    35 => array('surname35', 'forename35', 'initials35', 'first35', 'last35', 'author35'),
-    36 => array('surname36', 'forename36', 'initials36', 'first36', 'last36', 'author36'),
-    37 => array('surname37', 'forename37', 'initials37', 'first37', 'last37', 'author37'),
-    38 => array('surname38', 'forename38', 'initials38', 'first38', 'last38', 'author38'),
-    39 => array('surname39', 'forename39', 'initials39', 'first39', 'last39', 'author39'),
-    40 => array('surname40', 'forename40', 'initials40', 'first40', 'last40', 'author40'),
-    41 => array('surname41', 'forename41', 'initials41', 'first41', 'last41', 'author41'),
-    42 => array('surname42', 'forename42', 'initials42', 'first42', 'last42', 'author42'),
-    43 => array('surname43', 'forename43', 'initials43', 'first43', 'last43', 'author43'),
-    44 => array('surname44', 'forename44', 'initials44', 'first44', 'last44', 'author44'),
-    45 => array('surname45', 'forename45', 'initials45', 'first45', 'last45', 'author45'),
-    46 => array('surname46', 'forename46', 'initials46', 'first46', 'last46', 'author46'),
-    47 => array('surname47', 'forename47', 'initials47', 'first47', 'last47', 'author47'),
-    48 => array('surname48', 'forename48', 'initials48', 'first48', 'last48', 'author48'),
-    49 => array('surname49', 'forename49', 'initials49', 'first49', 'last49', 'author49'),
-    50 => array('surname50', 'forename50', 'initials50', 'first50', 'last50', 'author50'),
-    51 => array('surname51', 'forename51', 'initials51', 'first51', 'last51', 'author51'),
-    52 => array('surname52', 'forename52', 'initials52', 'first52', 'last52', 'author52'),
-    53 => array('surname53', 'forename53', 'initials53', 'first53', 'last53', 'author53'),
-    54 => array('surname54', 'forename54', 'initials54', 'first54', 'last54', 'author54'),
-    55 => array('surname55', 'forename55', 'initials55', 'first55', 'last55', 'author55'),
-    56 => array('surname56', 'forename56', 'initials56', 'first56', 'last56', 'author56'),
-    57 => array('surname57', 'forename57', 'initials57', 'first57', 'last57', 'author57'),
-    58 => array('surname58', 'forename58', 'initials58', 'first58', 'last58', 'author58'),
-    59 => array('surname59', 'forename59', 'initials59', 'first59', 'last59', 'author59'),
-    60 => array('surname60', 'forename60', 'initials60', 'first60', 'last60', 'author60'),
-    61 => array('surname61', 'forename61', 'initials61', 'first61', 'last61', 'author61'),
-    62 => array('surname62', 'forename62', 'initials62', 'first62', 'last62', 'author62'),
-    63 => array('surname63', 'forename63', 'initials63', 'first63', 'last63', 'author63'),
-    64 => array('surname64', 'forename64', 'initials64', 'first64', 'last64', 'author64'),
-    65 => array('surname65', 'forename65', 'initials65', 'first65', 'last65', 'author65'),
-    66 => array('surname66', 'forename66', 'initials66', 'first66', 'last66', 'author66'),
-    67 => array('surname67', 'forename67', 'initials67', 'first67', 'last67', 'author67'),
-    68 => array('surname68', 'forename68', 'initials68', 'first68', 'last68', 'author68'),
-    69 => array('surname69', 'forename69', 'initials69', 'first69', 'last69', 'author69'),
-    70 => array('surname70', 'forename70', 'initials70', 'first70', 'last70', 'author70'),
-    71 => array('surname71', 'forename71', 'initials71', 'first71', 'last71', 'author71'),
-    72 => array('surname72', 'forename72', 'initials72', 'first72', 'last72', 'author72'),
-    73 => array('surname73', 'forename73', 'initials73', 'first73', 'last73', 'author73'),
-    74 => array('surname74', 'forename74', 'initials74', 'first74', 'last74', 'author74'),
-    75 => array('surname75', 'forename75', 'initials75', 'first75', 'last75', 'author75'),
-    76 => array('surname76', 'forename76', 'initials76', 'first76', 'last76', 'author76'),
-    77 => array('surname77', 'forename77', 'initials77', 'first77', 'last77', 'author77'),
-    78 => array('surname78', 'forename78', 'initials78', 'first78', 'last78', 'author78'),
-    79 => array('surname79', 'forename79', 'initials79', 'first79', 'last79', 'author79'),
-    80 => array('surname80', 'forename80', 'initials80', 'first80', 'last80', 'author80'),
-    81 => array('surname81', 'forename81', 'initials81', 'first81', 'last81', 'author81'),
-    82 => array('surname82', 'forename82', 'initials82', 'first82', 'last82', 'author82'),
-    83 => array('surname83', 'forename83', 'initials83', 'first83', 'last83', 'author83'),
-    84 => array('surname84', 'forename84', 'initials84', 'first84', 'last84', 'author84'),
-    85 => array('surname85', 'forename85', 'initials85', 'first85', 'last85', 'author85'),
-    86 => array('surname86', 'forename86', 'initials86', 'first86', 'last86', 'author86'),
-    87 => array('surname87', 'forename87', 'initials87', 'first87', 'last87', 'author87'),
-    88 => array('surname88', 'forename88', 'initials88', 'first88', 'last88', 'author88'),
-    89 => array('surname89', 'forename89', 'initials89', 'first89', 'last89', 'author89'),
-    90 => array('surname90', 'forename90', 'initials90', 'first90', 'last90', 'author90'),
-    91 => array('surname91', 'forename91', 'initials91', 'first91', 'last91', 'author91'),
-    92 => array('surname92', 'forename92', 'initials92', 'first92', 'last92', 'author92'),
-    93 => array('surname93', 'forename93', 'initials93', 'first93', 'last93', 'author93'),
-    94 => array('surname94', 'forename94', 'initials94', 'first94', 'last94', 'author94'),
-    95 => array('surname95', 'forename95', 'initials95', 'first95', 'last95', 'author95'),
-    96 => array('surname96', 'forename96', 'initials96', 'first96', 'last96', 'author96'),
-    97 => array('surname97', 'forename97', 'initials97', 'first97', 'last97', 'author97'),
-    98 => array('surname98', 'forename98', 'initials98', 'first98', 'last98', 'author98'),
-    99 => array('surname99', 'forename99', 'initials99', 'first99', 'last99', 'author99'),
-);
-
-const FLATTENED_AUTHOR_PARAMETERS = array('surname', 'forename', 'initials', 
-    'first'  , 'last'  , 'author',
-    'surname1' , 'forename1' , 'initials1' , 'first1' , 'last1' , 'author1' , 'authors', 'vauthors',
-    'surname2' , 'forename2' , 'initials2' , 'first2' , 'last2' , 'author2' , 'coauthors', 'coauthor',
-    'surname3' , 'forename3' , 'initials3' , 'first3' , 'last3' , 'author3' ,
-    'surname4' , 'forename4' , 'initials4' , 'first4' , 'last4' , 'author4' ,
-    'surname5' , 'forename5' , 'initials5' , 'first5' , 'last5' , 'author5' ,
-    'surname6' , 'forename6' , 'initials6' , 'first6' , 'last6' , 'author6' ,
-    'surname7' , 'forename7' , 'initials7' , 'first7' , 'last7' , 'author7' ,
-    'surname8' , 'forename8' , 'initials8' , 'first8' , 'last8' , 'author8' ,
-    'surname9' , 'forename9' , 'initials9' , 'first9' , 'last9' , 'author9' ,
-    'surname10', 'forename10', 'initials10', 'first10', 'last10', 'author10',
-    'surname11', 'forename11', 'initials11', 'first11', 'last11', 'author11',
-    'surname12', 'forename12', 'initials12', 'first12', 'last12', 'author12',
-    'surname13', 'forename13', 'initials13', 'first13', 'last13', 'author13',
-    'surname14', 'forename14', 'initials14', 'first14', 'last14', 'author14',
-    'surname15', 'forename15', 'initials15', 'first15', 'last15', 'author15',
-    'surname16', 'forename16', 'initials16', 'first16', 'last16', 'author16',
-    'surname17', 'forename17', 'initials17', 'first17', 'last17', 'author17',
-    'surname18', 'forename18', 'initials18', 'first18', 'last18', 'author18',
-    'surname19', 'forename19', 'initials19', 'first19', 'last19', 'author19',
-    'surname20', 'forename20', 'initials20', 'first20', 'last20', 'author20',
-    'surname21', 'forename21', 'initials21', 'first21', 'last21', 'author21',
-    'surname22', 'forename22', 'initials22', 'first22', 'last22', 'author22',
-    'surname23', 'forename23', 'initials23', 'first23', 'last23', 'author23',
-    'surname24', 'forename24', 'initials24', 'first24', 'last24', 'author24',
-    'surname25', 'forename25', 'initials25', 'first25', 'last25', 'author25',
-    'surname26', 'forename26', 'initials26', 'first26', 'last26', 'author26',
-    'surname27', 'forename27', 'initials27', 'first27', 'last27', 'author27',
-    'surname28', 'forename28', 'initials28', 'first28', 'last28', 'author28',
-    'surname29', 'forename29', 'initials29', 'first29', 'last29', 'author29',
-    'surname30', 'forename30', 'initials30', 'first30', 'last30', 'author30',
-    'surname31', 'forename31', 'initials31', 'first31', 'last31', 'author31',
-    'surname32', 'forename32', 'initials32', 'first32', 'last32', 'author32',
-    'surname33', 'forename33', 'initials33', 'first33', 'last33', 'author33',
-    'surname34', 'forename34', 'initials34', 'first34', 'last34', 'author34',
-    'surname35', 'forename35', 'initials35', 'first35', 'last35', 'author35',
-    'surname36', 'forename36', 'initials36', 'first36', 'last36', 'author36',
-    'surname37', 'forename37', 'initials37', 'first37', 'last37', 'author37',
-    'surname38', 'forename38', 'initials38', 'first38', 'last38', 'author38',
-    'surname39', 'forename39', 'initials39', 'first39', 'last39', 'author39',
-    'surname40', 'forename40', 'initials40', 'first40', 'last40', 'author40',
-    'surname41', 'forename41', 'initials41', 'first41', 'last41', 'author41',
-    'surname42', 'forename42', 'initials42', 'first42', 'last42', 'author42',
-    'surname43', 'forename43', 'initials43', 'first43', 'last43', 'author43',
-    'surname44', 'forename44', 'initials44', 'first44', 'last44', 'author44',
-    'surname45', 'forename45', 'initials45', 'first45', 'last45', 'author45',
-    'surname46', 'forename46', 'initials46', 'first46', 'last46', 'author46',
-    'surname47', 'forename47', 'initials47', 'first47', 'last47', 'author47',
-    'surname48', 'forename48', 'initials48', 'first48', 'last48', 'author48',
-    'surname49', 'forename49', 'initials49', 'first49', 'last49', 'author49',
-    'surname50', 'forename50', 'initials50', 'first50', 'last50', 'author50',
-    'surname51', 'forename51', 'initials51', 'first51', 'last51', 'author51',
-    'surname52', 'forename52', 'initials52', 'first52', 'last52', 'author52',
-    'surname53', 'forename53', 'initials53', 'first53', 'last53', 'author53',
-    'surname54', 'forename54', 'initials54', 'first54', 'last54', 'author54',
-    'surname55', 'forename55', 'initials55', 'first55', 'last55', 'author55',
-    'surname56', 'forename56', 'initials56', 'first56', 'last56', 'author56',
-    'surname57', 'forename57', 'initials57', 'first57', 'last57', 'author57',
-    'surname58', 'forename58', 'initials58', 'first58', 'last58', 'author58',
-    'surname59', 'forename59', 'initials59', 'first59', 'last59', 'author59',
-    'surname60', 'forename60', 'initials60', 'first60', 'last60', 'author60',
-    'surname61', 'forename61', 'initials61', 'first61', 'last61', 'author61',
-    'surname62', 'forename62', 'initials62', 'first62', 'last62', 'author62',
-    'surname63', 'forename63', 'initials63', 'first63', 'last63', 'author63',
-    'surname64', 'forename64', 'initials64', 'first64', 'last64', 'author64',
-    'surname65', 'forename65', 'initials65', 'first65', 'last65', 'author65',
-    'surname66', 'forename66', 'initials66', 'first66', 'last66', 'author66',
-    'surname67', 'forename67', 'initials67', 'first67', 'last67', 'author67',
-    'surname68', 'forename68', 'initials68', 'first68', 'last68', 'author68',
-    'surname69', 'forename69', 'initials69', 'first69', 'last69', 'author69',
-    'surname70', 'forename70', 'initials70', 'first70', 'last70', 'author70',
-    'surname71', 'forename71', 'initials71', 'first71', 'last71', 'author71',
-    'surname72', 'forename72', 'initials72', 'first72', 'last72', 'author72',
-    'surname73', 'forename73', 'initials73', 'first73', 'last73', 'author73',
-    'surname74', 'forename74', 'initials74', 'first74', 'last74', 'author74',
-    'surname75', 'forename75', 'initials75', 'first75', 'last75', 'author75',
-    'surname76', 'forename76', 'initials76', 'first76', 'last76', 'author76',
-    'surname77', 'forename77', 'initials77', 'first77', 'last77', 'author77',
-    'surname78', 'forename78', 'initials78', 'first78', 'last78', 'author78',
-    'surname79', 'forename79', 'initials79', 'first79', 'last79', 'author79',
-    'surname80', 'forename80', 'initials80', 'first80', 'last80', 'author80',
-    'surname81', 'forename81', 'initials81', 'first81', 'last81', 'author81',
-    'surname82', 'forename82', 'initials82', 'first82', 'last82', 'author82',
-    'surname83', 'forename83', 'initials83', 'first83', 'last83', 'author83',
-    'surname84', 'forename84', 'initials84', 'first84', 'last84', 'author84',
-    'surname85', 'forename85', 'initials85', 'first85', 'last85', 'author85',
-    'surname86', 'forename86', 'initials86', 'first86', 'last86', 'author86',
-    'surname87', 'forename87', 'initials87', 'first87', 'last87', 'author87',
-    'surname88', 'forename88', 'initials88', 'first88', 'last88', 'author88',
-    'surname89', 'forename89', 'initials89', 'first89', 'last89', 'author89',
-    'surname90', 'forename90', 'initials90', 'first90', 'last90', 'author90',
-    'surname91', 'forename91', 'initials91', 'first91', 'last91', 'author91',
-    'surname92', 'forename92', 'initials92', 'first92', 'last92', 'author92',
-    'surname93', 'forename93', 'initials93', 'first93', 'last93', 'author93',
-    'surname94', 'forename94', 'initials94', 'first94', 'last94', 'author94',
-    'surname95', 'forename95', 'initials95', 'first95', 'last95', 'author95',
-    'surname96', 'forename96', 'initials96', 'first96', 'last96', 'author96',
-    'surname97', 'forename97', 'initials97', 'first97', 'last97', 'author97',
-    'surname98', 'forename98', 'initials98', 'first98', 'last98', 'author98',
-    'surname99', 'forename99', 'initials99', 'first99', 'last99', 'author99',
-);
-
-
-const PARAMETER_LIST = array('ARXIV', 'ASIN', 'ASIN-TLD', 'BIBCODE', 'DOI', 'ID',
-'ISBN', 'ISBN13', 'ISSN', 'JFM', 'JSTOR', 'LCCN', 'MR', 'OCLC', 'OL', 'OSTI', 'PMC', 'PMID',
-'RFC', 'SSRN', 'URL', 'ZBL', 'Author', 'DoiBroken', 'Editor', 'EditorGiven', 'EditorSurname',
- 'Embargo', 'PPPrefix', 'PPrefix', 'Ref', 'Author1', 'Author2', 'Author3', 'Author4', 'Author5',
- 'Author6', 'Author7', 'Author8', 'Author9', 'Author10', 'Author11', 'Author12', 'Author13',
- 'Author14', 'Author15', 'Author16', 'Author17', 'Author18', 'Author19', 'Author20', 'Author21', 
- 'Author22', 'Author23', 'Author24', 'Author25', 'Author26', 'Author27', 'Author28', 'Author29', 
- 'Author30', 'Author31', 'Author32', 'Author33', 'Author34', 'Author35', 'Author36', 'Author37',
- 'Author38', 'Author39', 'Author40', 'Author41', 'Author42', 'Author43', 'Author44', 'Author45',
- 'Author46', 'Author47', 'Author48', 'Author49', 'Author50', 'Author51', 'Author52', 'Author53',
- 'Author54', 'Author55', 'Author56', 'Author57', 'Author58', 'Author59', 'Author60', 'Author61',
- 'Author62', 'Author63', 'Author64', 'Author65', 'Author66', 'Author67', 'Author68', 'Author69',
- 'Author70', 'Author71', 'Author72', 'Author73', 'Author74', 'Author75', 'Author76', 'Author77',
- 'Author78', 'Author79', 'Author80', 'Author81', 'Author82', 'Author83', 'Author84', 'Author85',
- 'Author86', 'Author87', 'Author88', 'Author89', 'Author90', 'Author91', 'Author92', 'Author93',
- 'Author94', 'Author95', 'Author96', 'Author97', 'Author98', 'Author99', 'Editor1', 'Editor2',
- 'Editor3', 'Editor4', 'Editor5', 'Editor6', 'Editor7', 'Editor8', 'Editor9', 'Editor10',
- 'Editor11', 'Editor12', 'Editor13', 'Editor14', 'Editor15', 'Editor16', 'Editor17', 'Editor18',
- 'Editor19', 'Editor20', 'Editor21', 'Editor22', 'Editor23', 'Editor24', 'Editor25', 'Editor26',
- 'Editor27', 'Editor28', 'Editor29', 'Editor30', 'Editor31', 'Editor32', 'Editor33', 'Editor34',
- 'Editor35', 'Editor36', 'Editor37', 'Editor38', 'Editor39', 'Editor40', 'Editor41', 'Editor42',
- 'Editor43', 'Editor44', 'Editor45', 'EditorGiven1', 'EditorGiven2', 'EditorGiven3', 'EditorGiven4',
- 'EditorGiven5', 'EditorGiven6', 'EditorGiven7', 'EditorGiven8', 'EditorGiven9', 'EditorGiven10',
- 'EditorGiven11', 'EditorGiven12', 'EditorGiven13', 'EditorGiven14', 'EditorGiven15',
- 'EditorGiven16', 'EditorGiven17', 'EditorGiven18', 'EditorGiven19', 'EditorGiven20',
- 'EditorGiven21', 'EditorGiven22', 'EditorGiven23', 'EditorGiven24', 'EditorGiven25',
- 'EditorGiven26', 'EditorGiven27', 'EditorGiven28', 'EditorGiven29', 'EditorGiven30',
- 'EditorGiven31', 'EditorGiven32', 'EditorGiven33', 'EditorGiven34', 'EditorGiven35',
- 'EditorGiven36', 'EditorGiven37', 'EditorGiven38', 'EditorGiven39', 'EditorGiven40',
- 'EditorGiven41', 'EditorGiven42', 'EditorGiven43', 'EditorGiven44', 'EditorGiven45',
- 'EditorSurname1', 'EditorSurname2', 'EditorSurname3', 'EditorSurname4', 'EditorSurname5',
- 'EditorSurname6', 'EditorSurname7', 'EditorSurname8', 'EditorSurname9', 'EditorSurname10',
- 'EditorSurname11', 'EditorSurname12', 'EditorSurname13', 'EditorSurname14', 'EditorSurname15',
- 'EditorSurname16', 'EditorSurname17', 'EditorSurname18', 'EditorSurname19', 'EditorSurname20',
- 'EditorSurname21', 'EditorSurname22', 'EditorSurname23', 'EditorSurname24', 'EditorSurname25',
- 'EditorSurname26', 'EditorSurname27', 'EditorSurname28', 'EditorSurname29', 'EditorSurname30',
- 'EditorSurname31', 'EditorSurname32', 'EditorSurname33', 'EditorSurname34', 'EditorSurname35',
- 'EditorSurname36', 'EditorSurname37', 'EditorSurname38', 'EditorSurname39', 'EditorSurname40',
- 'EditorSurname41', 'EditorSurname42', 'EditorSurname43', 'EditorSurname44', 'EditorSurname45',
- 'accessdate', 'agency', 'airdate', 'archivedate', 'archive-date', 'archiveurl', 'archive-url',
- 'article', 'arxiv', 'asin', 'asin-tld', 'at', 'author', 'author-first', 'authorformat',
- 'author-format', 'author-last', 'authorlink', 'author-link', 'authormask', 'author-mask',
- 'author-name-separator', 'authors', 'author-separator', 'bibcode', 'booktitle', 'callsign',
- 'cartography', 'chapter', 'chapterlink', 'chapterurl', 'chapter-url', 'city', 'class', 'conference',
- 'conferenceurl', 'conference-url', 'contribution', 'contributionurl', 'contribution-url', 'date',
- 'deadurl', 'degree', 'department', 'dictionary', 'displayauthors', 'display-authors',
- 'displayeditors', 'display-editors', 'docket', 'doi', 'doi_brokendate', 'doi_inactivedate',
- 'doi-broken-date',
- 'edition', 'editor', 'editor-first', 'editorformat', 'editor-format', 'editor-last',
- 'editorlink', 'editor-link', 'editormask', 'editor-mask', 'editor-name-separator',
- 'editors', 'editor-separator', 'embargo', 'encyclopaedia', 'encyclopedia', 'entry',
- 'episodelink', 'eprint', 'event', 'eventurl', 'first', 'format', 'given', 'host', 'id',
- 'ignoreisbnerror', 'ignore-isbn-error', 'in', 'inset', 'institution', 'interviewer',
- 'isbn', 'isbn13', 'issn', 'issue', 'jfm', 'journal', 'jstor', 'language', 'last',
- 'lastauthoramp', 'laydate', 'laysource', 'laysummary', 'layurl', 'lccn', 'location',
- 'magazine', 'medium', 'minutes', 'mr', 'name-separator', 'network', 'newspaper', 'nocat',
- 'nopp', 'notracking', 'no-tracking', 'number', 'oclc', 'ol', 'origyear', 'osti', 'others',
- 'p', 'page', 'pages', 'people', 'periodical', 'place', 'pmc', 'pmc-embargo-date', 'pmid',
- 'postscript', 'pp', 'program', 'publicationdate', 'publication-date', 'publicationplace',
- 'publication-place', 'publisher', 'quotation', 'quote', 'ref', 'registration', 'rfc',
- 'scale', 'season', 'section', 'sectionurl', 'separator', 'series', 'serieslink', 'seriesno',
- 'seriesnumber', 'series-separator', 'ssrn', 'station', 'subject', 'subjectlink', 'subscription',
- 'surname', 'template doc demo', 'time', 'timecaption', 'title', 'titlelink', 'trans_chapter',
- 'trans_title', 'trans-chapter', 'transcript', 'transcripturl', 'transcript-url',
- 'trans-title', 'type', 'url', 'version', 'via', 'volume', 'website', 'work', 'year', 'zbl',
- 'author1', 'author1-first', 'author1-last', 'author1link', 'author1-link', 'author1mask',
- 'author1-mask', 'author2', 'author2-first', 'author2-last', 'author2link', 'author2-link',
- 'author2mask', 'author2-mask', 'author3', 'author3-first', 'author3-last', 'author3link',
- 'author3-link', 'author3mask', 'author3-mask', 'author4', 'author4-first', 'author4-last',
- 'author4link', 'author4-link', 'author4mask', 'author4-mask', 'author5', 'author5-first',
- 'author5-last', 'author5link', 'author5-link', 'author5mask', 'author5-mask', 'author6',
- 'author6-first', 'author6-last', 'author6link', 'author6-link', 'author6mask', 'author6-mask',
- 'author7', 'author7-first', 'author7-last', 'author7link', 'author7-link', 'author7mask',
- 'author7-mask', 'author8', 'author8-first', 'author8-last', 'author8link', 'author8-link',
- 'author8mask', 'author8-mask', 'author9', 'author9-first', 'author9-last', 'author9link', 
- 'author9-link', 'author9mask', 'author9-mask', 'author10', 'author10-first', 'author10-last',
- 'author10link', 'author10-link', 'author10mask', 'author10-mask', 'author11', 'author11-first',
- 'author11-last', 'author11link', 'author11-link', 'author11mask', 'author11-mask', 'author12',
- 'author12-first', 'author12-last', 'author12link', 'author12-link', 'author12mask',
- 'author12-mask', 'author13', 'author13-first', 'author13-last', 'author13link', 'author13-link',
- 'author13mask', 'author13-mask', 'author14', 'author14-first', 'author14-last', 'author14link',
- 'author14-link', 'author14mask', 'author14-mask', 'author15', 'author15-first', 'author15-last',
- 'author15link', 'author15-link', 'author15mask', 'author15-mask', 'author16', 'author16-first',
- 'author16-last', 'author16link', 'author16-link', 'author16mask', 'author16-mask', 'author17',
- 'author17-first', 'author17-last', 'author17link', 'author17-link', 'author17mask', 'author17-mask', 
- 'author18', 'author18-first', 'author18-last', 'author18link', 'author18-link', 'author18mask',
- 'author18-mask', 'author19', 'author19-first', 'author19-last', 'author19link', 'author19-link',
- 'author19mask', 'author19-mask', 'author20', 'author20-first', 'author20-last', 'author20link',
- 'author20-link', 'author20mask', 'author20-mask', 'author21', 'author21-first', 'author21-last',
- 'author21link', 'author21-link', 'author21mask', 'author21-mask', 'author22', 'author22-first',
- 'author22-last', 'author22link', 'author22-link', 'author22mask', 'author22-mask', 'author23',
- 'author23-first', 'author23-last', 'author23link', 'author23-link', 'author23mask', 'author23-mask',
- 'author24', 'author24-first', 'author24-last', 'author24link', 'author24-link', 'author24mask',
- 'author24-mask', 'author25', 'author25-first', 'author25-last', 'author25link', 'author25-link',
- 'author25mask', 'author25-mask', 'author26', 'author26-first', 'author26-last', 'author26link',
- 'author26-link', 'author26mask', 'author26-mask', 'author27', 'author27-first', 'author27-last',
- 'author27link', 'author27-link', 'author27mask', 'author27-mask', 'author28', 'author28-first',
- 'author28-last', 'author28link', 'author28-link', 'author28mask', 'author28-mask', 'author29',
- 'author29-first', 'author29-last', 'author29link', 'author29-link', 'author29mask', 'author29-mask',
- 'author30', 'author30-first', 'author30-last', 'author30link', 'author30-link', 'author30mask',
- 'author30-mask', 'author31', 'author31-first', 'author31-last', 'author31link', 'author31-link',
- 'author31mask', 'author31-mask', 'author32', 'author32-first', 'author32-last', 'author32link',
- 'author32-link', 'author32mask', 'author32-mask', 'author33', 'author33-first', 'author33-last',
- 'author33link', 'author33-link', 'author33mask', 'author33-mask', 'author34', 'author34-first',
- 'author34-last', 'author34link', 'author34-link', 'author34mask', 'author34-mask', 'author35',
- 'author35-first', 'author35-last', 'author35link', 'author35-link', 'author35mask', 'author35-mask',
- 'author36', 'author36-first', 'author36-last', 'author36link', 'author36-link', 'author36mask',
- 'author36-mask', 'author37', 'author37-first', 'author37-last', 'author37link', 'author37-link',
- 'author37mask', 'author37-mask', 'author38', 'author38-first', 'author38-last', 'author38link',
- 'author38-link', 'author38mask', 'author38-mask', 'author39', 'author39-first', 'author39-last',
- 'author39link', 'author39-link', 'author39mask', 'author39-mask', 'author40', 'author40-first',
- 'author40-last', 'author40link', 'author40-link', 'author40mask', 'author40-mask', 'author41',
- 'author41-first', 'author41-last', 'author41link', 'author41-link', 'author41mask', 'author41-mask',
- 'author42', 'author42-first', 'author42-last', 'author42link', 'author42-link', 'author42mask', 
- 'author42-mask', 'author43', 'author43-first', 'author43-last', 'author43link', 'author43-link',
- 'author43mask', 'author43-mask', 'author44', 'author44-first', 'author44-last', 'author44link',
- 'author44-link', 'author44mask', 'author44-mask', 'author45', 'author45-first', 'author45-last',
- 'author45link', 'author45-link', 'author45mask', 'author45-mask', 'author46', 'author46-first',
- 'author46-last', 'author46link', 'author46-link', 'author46mask', 'author46-mask', 'author47', 
- 'author47-first', 'author47-last', 'author47link', 'author47-link', 'author47mask',
- 'author47-mask', 'author48', 'author48-first', 'author48-last', 'author48link', 
- 'author48-link', 'author48mask', 'author48-mask', 'author49', 'author49-first', 
- 'author49-last', 'author49link', 'author49-link', 'author49mask', 'author49-mask',
- 'author50', 'author50-first', 'author50-last', 'author50link', 'author50-link', 'author50mask', 
- 'author50-mask', 'author51', 'author51-first', 'author51-last', 'author51link', 'author51-link',
- 'author51mask', 'author51-mask', 'author52', 'author52-first', 'author52-last', 'author52link',
- 'author52-link', 'author52mask', 'author52-mask', 'author53', 'author53-first', 'author53-last',
- 'author53link', 'author53-link', 'author53mask', 'author53-mask', 'author54', 'author54-first',
- 'author54-last', 'author54link', 'author54-link', 'author54mask', 'author54-mask', 'author55',
- 'author55-first', 'author55-last', 'author55link', 'author55-link', 'author55mask', 'author55-mask',
- 'author56', 'author56-first', 'author56-last', 'author56link', 'author56-link', 'author56mask',
- 'author56-mask', 'author57', 'author57-first', 'author57-last', 'author57link', 'author57-link',
- 'author57mask', 'author57-mask', 'author58', 'author58-first', 'author58-last', 'author58link',
- 'author58-link', 'author58mask', 'author58-mask', 'author59', 'author59-first', 'author59-last',
- 'author59link', 'author59-link', 'author59mask', 'author59-mask', 'author60', 'author60-first',
- 'author60-last', 'author60link', 'author60-link', 'author60mask', 'author60-mask', 'author61',
- 'author61-first', 'author61-last', 'author61link', 'author61-link', 'author61mask', 'author61-mask', 
-'author62', 'author62-first', 'author62-last', 'author62link', 'author62-link',
- 'author62mask', 'author62-mask', 'author63', 'author63-first', 'author63-last',
- 'author63link', 'author63-link', 'author63mask', 'author63-mask', 'author64',
- 'author64-first', 'author64-last', 'author64link', 'author64-link', 'author64mask',
- 'author64-mask', 'author65', 'author65-first', 'author65-last', 'author65link',
- 'author65-link', 'author65mask', 'author65-mask', 'author66', 'author66-first',
- 'author66-last', 'author66link', 'author66-link', 'author66mask', 'author66-mask',
- 'author67', 'author67-first', 'author67-last', 'author67link', 'author67-link',
- 'author67mask', 'author67-mask', 'author68', 'author68-first', 'author68-last',
- 'author68link', 'author68-link', 'author68mask', 'author68-mask', 'author69',
- 'author69-first', 'author69-last', 'author69link', 'author69-link', 'author69mask',
- 'author69-mask', 'author70', 'author70-first', 'author70-last', 'author70link',
- 'author70-link', 'author70mask', 'author70-mask', 'author71', 'author71-first',
- 'author71-last', 'author71link', 'author71-link', 'author71mask', 'author71-mask',
- 'author72', 'author72-first', 'author72-last', 'author72link', 'author72-link',
- 'author72mask', 'author72-mask', 'author73', 'author73-first', 'author73-last',
- 'author73link', 'author73-link', 'author73mask', 'author73-mask', 'author74',
- 'author74-first', 'author74-last', 'author74link', 'author74-link', 'author74mask',
- 'author74-mask', 'author75', 'author75-first', 'author75-last', 'author75link',
- 'author75-link', 'author75mask', 'author75-mask', 'author76', 'author76-first',
- 'author76-last', 'author76link', 'author76-link', 'author76mask', 'author76-mask',
- 'author77', 'author77-first', 'author77-last', 'author77link', 'author77-link',
- 'author77mask', 'author77-mask', 'author78', 'author78-first', 'author78-last',
- 'author78link', 'author78-link', 'author78mask', 'author78-mask', 'author79',
- 'author79-first', 'author79-last', 'author79link', 'author79-link', 'author79mask',
- 'author79-mask', 'author80', 'author80-first', 'author80-last', 'author80link',
- 'author80-link', 'author80mask', 'author80-mask', 'author81', 'author81-first',
- 'author81-last', 'author81link', 'author81-link', 'author81mask', 'author81-mask',
- 'author82', 'author82-first', 'author82-last', 'author82link', 'author82-link',
- 'author82mask', 'author82-mask', 'author83', 'author83-first', 'author83-last',
- 'author83link', 'author83-link', 'author83mask', 'author83-mask', 'author84',
- 'author84-first', 'author84-last', 'author84link', 'author84-link', 'author84mask',
- 'author84-mask', 'author85', 'author85-first', 'author85-last', 'author85link',
- 'author85-link', 'author85mask', 'author85-mask', 'author86', 'author86-first',
- 'author86-last', 'author86link', 'author86-link', 'author86mask', 'author86-mask',
- 'author87', 'author87-first', 'author87-last', 'author87link', 'author87-link',
- 'author87mask', 'author87-mask', 'author88', 'author88-first', 'author88-last',
- 'author88link', 'author88-link', 'author88mask', 'author88-mask', 'author89',
- 'author89-first', 'author89-last', 'author89link', 'author89-link', 'author89mask',
- 'author89-mask', 'author90', 'author90-first', 'author90-last', 'author90link',
- 'author90-link', 'author90mask', 'author90-mask', 'author91', 'author91-first',
- 'author91-last', 'author91link', 'author91-link', 'author91mask', 'author91-mask',
- 'author92', 'author92-first', 'author92-last', 'author92link', 'author92-link',
- 'author92mask', 'author92-mask', 'author93', 'author93-first', 'author93-last',
- 'author93link', 'author93-link', 'author93mask', 'author93-mask', 'author94',
- 'author94-first', 'author94-last', 'author94link', 'author94-link', 'author94mask',
- 'author94-mask', 'author95', 'author95-first', 'author95-last', 'author95link',
- 'author95-link', 'author95mask', 'author95-mask', 'author96', 'author96-first',
- 'author96-last', 'author96link', 'author96-link', 'author96mask', 'author96-mask',
- 'author97', 'author97-first', 'author97-last', 'author97link', 'author97-link',
- 'author97mask', 'author97-mask', 'author98', 'author98-first', 'author98-last',
- 'author98link', 'author98-link', 'author98mask', 'author98-mask', 'author99',
- 'author99-first', 'author99-last', 'author99link', 'author99-link', 'author99mask',
- 'author99-mask', 'author-first1', 'author-first2', 'author-first3', 'author-first4',
- 'author-first5', 'author-first6', 'author-first7', 'author-first8', 'author-first9',
- 'author-first10', 'author-first11', 'author-first12', 'author-first13', 'author-first14',
- 'author-first15', 'author-first16', 'author-first17', 'author-first18', 'author-first19',
- 'author-first20', 'author-first21', 'author-first22', 'author-first23', 'author-first24',
- 'author-first25', 'author-first26', 'author-first27', 'author-first28', 'author-first29',
- 'author-first30', 'author-first31', 'author-first32', 'author-first33', 'author-first34',
- 'author-first35', 'author-first36', 'author-first37', 'author-first38', 'author-first39',
- 'author-first40', 'author-first41', 'author-first42', 'author-first43', 'author-first44',
- 'author-first45', 'author-first46', 'author-first47', 'author-first48', 'author-first49',
- 'author-first50', 'author-first51', 'author-first52', 'author-first53', 'author-first54',
- 'author-first55', 'author-first56', 'author-first57', 'author-first58', 'author-first59',
- 'author-first60', 'author-first61', 'author-first62', 'author-first63', 'author-first64',
- 'author-first65', 'author-first66', 'author-first67', 'author-first68', 'author-first69',
- 'author-first70', 'author-first71', 'author-first72', 'author-first73', 'author-first74',
- 'author-first75', 'author-first76', 'author-first77', 'author-first78', 'author-first79',
- 'author-first80', 'author-first81', 'author-first82', 'author-first83', 'author-first84',
- 'author-first85', 'author-first86', 'author-first87', 'author-first88', 'author-first89',
- 'author-first90', 'author-first91', 'author-first92', 'author-first93', 'author-first94',
- 'author-first95', 'author-first96', 'author-first97', 'author-first98', 'author-first99',
- 'author-last1', 'author-last2', 'author-last3', 'author-last4', 'author-last5',
- 'author-last6', 'author-last7', 'author-last8', 'author-last9', 'author-last10',
- 'author-last11', 'author-last12', 'author-last13', 'author-last14', 'author-last15',
- 'author-last16', 'author-last17', 'author-last18', 'author-last19', 'author-last20',
- 'author-last21', 'author-last22', 'author-last23', 'author-last24', 'author-last25',
- 'author-last26', 'author-last27', 'author-last28', 'author-last29', 'author-last30',
- 'author-last31', 'author-last32', 'author-last33', 'author-last34', 'author-last35',
- 'author-last36', 'author-last37', 'author-last38', 'author-last39', 'author-last40',
- 'author-last41', 'author-last42', 'author-last43', 'author-last44', 'author-last45',
- 'author-last46', 'author-last47', 'author-last48', 'author-last49', 'author-last50',
- 'author-last51', 'author-last52', 'author-last53', 'author-last54', 'author-last55',
- 'author-last56', 'author-last57', 'author-last58', 'author-last59', 'author-last60',
- 'author-last61', 'author-last62', 'author-last63', 'author-last64', 'author-last65',
- 'author-last66', 'author-last67', 'author-last68', 'author-last69', 'author-last70',
- 'author-last71', 'author-last72', 'author-last73', 'author-last74', 'author-last75',
- 'author-last76', 'author-last77', 'author-last78', 'author-last79', 'author-last80',
- 'author-last81', 'author-last82', 'author-last83', 'author-last84', 'author-last85',
- 'author-last86', 'author-last87', 'author-last88', 'author-last89', 'author-last90',
- 'author-last91', 'author-last92', 'author-last93', 'author-last94', 'author-last95',
- 'author-last96', 'author-last97', 'author-last98', 'author-last99', 'authorlink1',
- 'authorlink2', 'authorlink3', 'authorlink4', 'authorlink5', 'authorlink6',
- 'authorlink7', 'authorlink8', 'authorlink9', 'authorlink10', 'authorlink11',
- 'authorlink12', 'authorlink13', 'authorlink14', 'authorlink15', 'authorlink16',
- 'authorlink17', 'authorlink18', 'authorlink19', 'authorlink20', 'authorlink21',
- 'authorlink22', 'authorlink23', 'authorlink24', 'authorlink25', 'authorlink26',
- 'authorlink27', 'authorlink28', 'authorlink29', 'authorlink30', 'authorlink31',
- 'authorlink32', 'authorlink33', 'authorlink34', 'authorlink35', 'authorlink36',
- 'authorlink37', 'authorlink38', 'authorlink39', 'authorlink40', 'authorlink41',
- 'authorlink42', 'authorlink43', 'authorlink44', 'authorlink45', 'authorlink46',
- 'authorlink47', 'authorlink48', 'authorlink49', 'authorlink50', 'authorlink51',
- 'authorlink52', 'authorlink53', 'authorlink54', 'authorlink55', 'authorlink56',
- 'authorlink57', 'authorlink58', 'authorlink59', 'authorlink60', 'authorlink61',
- 'authorlink62', 'authorlink63', 'authorlink64', 'authorlink65', 'authorlink66',
- 'authorlink67', 'authorlink68', 'authorlink69', 'authorlink70', 'authorlink71',
- 'authorlink72', 'authorlink73', 'authorlink74', 'authorlink75', 'authorlink76',
- 'authorlink77', 'authorlink78', 'authorlink79', 'authorlink80', 'authorlink81',
- 'authorlink82', 'authorlink83', 'authorlink84', 'authorlink85', 'authorlink86',
- 'authorlink87', 'authorlink88', 'authorlink89', 'authorlink90', 'authorlink91',
- 'authorlink92', 'authorlink93', 'authorlink94', 'authorlink95', 'authorlink96',
- 'authorlink97', 'authorlink98', 'authorlink99', 'author-link1', 'author-link2',
- 'author-link3', 'author-link4', 'author-link5', 'author-link6', 'author-link7',
- 'author-link8', 'author-link9', 'author-link10', 'author-link11', 'author-link12',
- 'author-link13', 'author-link14', 'author-link15', 'author-link16', 'author-link17',
- 'author-link18', 'author-link19', 'author-link20', 'author-link21', 'author-link22',
- 'author-link23', 'author-link24', 'author-link25', 'author-link26', 'author-link27',
- 'author-link28', 'author-link29', 'author-link30', 'author-link31', 'author-link32',
- 'author-link33', 'author-link34', 'author-link35', 'author-link36', 'author-link37',
- 'author-link38', 'author-link39', 'author-link40', 'author-link41', 'author-link42',
- 'author-link43', 'author-link44', 'author-link45', 'author-link46', 'author-link47',
- 'author-link48', 'author-link49', 'author-link50', 'author-link51', 'author-link52',
- 'author-link53', 'author-link54', 'author-link55', 'author-link56', 'author-link57',
- 'author-link58', 'author-link59', 'author-link60', 'author-link61', 'author-link62',
- 'author-link63', 'author-link64', 'author-link65', 'author-link66', 'author-link67',
- 'author-link68', 'author-link69', 'author-link70', 'author-link71', 'author-link72',
- 'author-link73', 'author-link74', 'author-link75', 'author-link76', 'author-link77',
- 'author-link78', 'author-link79', 'author-link80', 'author-link81', 'author-link82',
- 'author-link83', 'author-link84', 'author-link85', 'author-link86', 'author-link87',
- 'author-link88', 'author-link89', 'author-link90', 'author-link91', 'author-link92',
- 'author-link93', 'author-link94', 'author-link95', 'author-link96', 'author-link97',
- 'author-link98', 'author-link99', 'authormask1', 'authormask2', 'authormask3',
- 'authormask4', 'authormask5', 'authormask6', 'authormask7', 'authormask8',
- 'authormask9', 'authormask10', 'authormask11', 'authormask12', 'authormask13',
- 'authormask14', 'authormask15', 'authormask16', 'authormask17', 'authormask18',
- 'authormask19', 'authormask20', 'authormask21', 'authormask22', 'authormask23',
- 'authormask24', 'authormask25', 'authormask26', 'authormask27', 'authormask28',
- 'authormask29', 'authormask30', 'authormask31', 'authormask32', 'authormask33',
- 'authormask34', 'authormask35', 'authormask36', 'authormask37', 'authormask38',
- 'authormask39', 'authormask40', 'authormask41', 'authormask42', 'authormask43',
- 'authormask44', 'authormask45', 'authormask46', 'authormask47', 'authormask48',
- 'authormask49', 'authormask50', 'authormask51', 'authormask52', 'authormask53',
- 'authormask54', 'authormask55', 'authormask56', 'authormask57', 'authormask58',
- 'authormask59', 'authormask60', 'authormask61', 'authormask62', 'authormask63',
- 'authormask64', 'authormask65', 'authormask66', 'authormask67', 'authormask68',
- 'authormask69', 'authormask70', 'authormask71', 'authormask72', 'authormask73',
- 'authormask74', 'authormask75', 'authormask76', 'authormask77', 'authormask78',
- 'authormask79', 'authormask80', 'authormask81', 'authormask82', 'authormask83',
- 'authormask84', 'authormask85', 'authormask86', 'authormask87', 'authormask88',
- 'authormask89', 'authormask90', 'authormask91', 'authormask92', 'authormask93',
- 'authormask94', 'authormask95', 'authormask96', 'authormask97', 'authormask98',
- 'authormask99', 'author-mask1', 'author-mask2', 'author-mask3', 'author-mask4',
- 'author-mask5', 'author-mask6', 'author-mask7', 'author-mask8', 'author-mask9',
- 'author-mask10', 'author-mask11', 'author-mask12', 'author-mask13', 'author-mask14',
- 'author-mask15', 'author-mask16', 'author-mask17', 'author-mask18', 'author-mask19',
- 'author-mask20', 'author-mask21', 'author-mask22', 'author-mask23', 'author-mask24',
- 'author-mask25', 'author-mask26', 'author-mask27', 'author-mask28', 'author-mask29',
- 'author-mask30', 'author-mask31', 'author-mask32', 'author-mask33', 'author-mask34',
- 'author-mask35', 'author-mask36', 'author-mask37', 'author-mask38', 'author-mask39',
- 'author-mask40', 'author-mask41', 'author-mask42', 'author-mask43', 'author-mask44',
- 'author-mask45', 'author-mask46', 'author-mask47', 'author-mask48', 'author-mask49',
- 'author-mask50', 'author-mask51', 'author-mask52', 'author-mask53', 'author-mask54',
- 'author-mask55', 'author-mask56', 'author-mask57', 'author-mask58', 'author-mask59',
- 'author-mask60', 'author-mask61', 'author-mask62', 'author-mask63', 'author-mask64',
- 'author-mask65', 'author-mask66', 'author-mask67', 'author-mask68', 'author-mask69',
- 'author-mask70', 'author-mask71', 'author-mask72', 'author-mask73', 'author-mask74',
- 'author-mask75', 'author-mask76', 'author-mask77', 'author-mask78', 'author-mask79',
- 'author-mask80', 'author-mask81', 'author-mask82', 'author-mask83', 'author-mask84',
- 'author-mask85', 'author-mask86', 'author-mask87', 'author-mask88', 'author-mask89',
- 'author-mask90', 'author-mask91', 'author-mask92', 'author-mask93', 'author-mask94',
- 'author-mask95', 'author-mask96', 'author-mask97', 'author-mask98', 'author-mask99',
- 'authors1', 'authors2', 'authors3', 'authors4', 'authors5',
- 'authors6', 'authors7', 'authors8', 'authors9', 'authors10',
- 'authors11', 'authors12', 'authors13', 'authors14', 'authors15',
- 'authors16', 'authors17', 'authors18', 'authors19', 'authors20',
- 'authors21', 'authors22', 'authors23', 'authors24', 'authors25',
- 'authors26', 'authors27', 'authors28', 'authors29', 'authors30',
- 'authors31', 'authors32', 'authors33', 'authors34', 'authors35',
- 'authors36', 'authors37', 'authors38', 'authors39', 'authors40',
- 'authors41', 'authors42', 'authors43', 'authors44', 'authors45',
- 'authors46', 'authors47', 'authors48', 'authors49', 'authors50',
- 'authors51', 'authors52', 'authors53', 'authors54', 'authors55',
- 'authors56', 'authors57', 'authors58', 'authors59', 'authors60',
- 'authors61', 'authors62', 'authors63', 'authors64', 'authors65',
- 'authors66', 'authors67', 'authors68', 'authors69', 'authors70',
- 'authors71', 'authors72', 'authors73', 'authors74', 'authors75',
- 'authors76', 'authors77', 'authors78', 'authors79', 'authors80',
- 'authors81', 'authors82', 'authors83', 'authors84', 'authors85',
- 'authors86', 'authors87', 'authors88', 'authors89', 'authors90',
- 'authors91', 'authors92', 'authors93', 'authors94', 'authors95',
- 'authors96', 'authors97', 'authors98', 'authors99', 'editor1',
- 'editor1-first', 'editor1-last', 'editor1link', 'editor1-link', 'editor1mask',
- 'editor1-mask', 'editor2', 'editor2-first', 'editor2-last', 'editor2link',
- 'editor2-link', 'editor2mask', 'editor2-mask', 'editor3', 'editor3-first',
- 'editor3-last', 'editor3link', 'editor3-link', 'editor3mask', 'editor3-mask',
- 'editor4', 'editor4-first', 'editor4-last', 'editor4link', 'editor4-link',
- 'editor4mask', 'editor4-mask', 'editor5', 'editor5-first', 'editor5-last',
- 'editor5link', 'editor5-link', 'editor5mask', 'editor5-mask', 'editor6',
- 'editor6-first', 'editor6-last', 'editor6link', 'editor6-link', 'editor6mask',
- 'editor6-mask', 'editor7', 'editor7-first', 'editor7-last', 'editor7link',
- 'editor7-link', 'editor7mask', 'editor7-mask', 'editor8', 'editor8-first',
- 'editor8-last', 'editor8link', 'editor8-link', 'editor8mask', 'editor8-mask',
- 'editor9', 'editor9-first', 'editor9-last', 'editor9link', 'editor9-link',
- 'editor9mask', 'editor9-mask', 'editor10', 'editor10-first', 'editor10-last',
- 'editor10link', 'editor10-link', 'editor10mask', 'editor10-mask', 'editor11',
- 'editor11-first', 'editor11-last', 'editor11link', 'editor11-link', 'editor11mask',
- 'editor11-mask', 'editor12', 'editor12-first', 'editor12-last', 'editor12link',
- 'editor12-link', 'editor12mask', 'editor12-mask', 'editor13', 'editor13-first',
- 'editor13-last', 'editor13link', 'editor13-link', 'editor13mask', 'editor13-mask',
- 'editor14', 'editor14-first', 'editor14-last', 'editor14link', 'editor14-link',
- 'editor14mask', 'editor14-mask', 'editor15', 'editor15-first', 'editor15-last',
- 'editor15link', 'editor15-link', 'editor15mask', 'editor15-mask', 'editor16',
- 'editor16-first', 'editor16-last', 'editor16link', 'editor16-link', 'editor16mask',
- 'editor16-mask', 'editor17', 'editor17-first', 'editor17-last', 'editor17link',
- 'editor17-link', 'editor17mask', 'editor17-mask', 'editor18', 'editor18-first',
- 'editor18-last', 'editor18link', 'editor18-link', 'editor18mask', 'editor18-mask',
- 'editor19', 'editor19-first', 'editor19-last', 'editor19link', 'editor19-link',
- 'editor19mask', 'editor19-mask', 'editor20', 'editor20-first', 'editor20-last',
- 'editor20link', 'editor20-link', 'editor20mask', 'editor20-mask', 'editor21',
- 'editor21-first', 'editor21-last', 'editor21link', 'editor21-link', 'editor21mask',
- 'editor21-mask', 'editor22', 'editor22-first', 'editor22-last', 'editor22link',
- 'editor22-link', 'editor22mask', 'editor22-mask', 'editor23', 'editor23-first',
- 'editor23-last', 'editor23link', 'editor23-link', 'editor23mask', 'editor23-mask',
- 'editor24', 'editor24-first', 'editor24-last', 'editor24link', 'editor24-link',
- 'editor24mask', 'editor24-mask', 'editor25', 'editor25-first', 'editor25-last',
- 'editor25link', 'editor25-link', 'editor25mask', 'editor25-mask', 'editor26',
- 'editor26-first', 'editor26-last', 'editor26link', 'editor26-link', 'editor26mask',
- 'editor26-mask', 'editor27', 'editor27-first', 'editor27-last', 'editor27link',
- 'editor27-link', 'editor27mask', 'editor27-mask', 'editor28', 'editor28-first',
- 'editor28-last', 'editor28link', 'editor28-link', 'editor28mask', 'editor28-mask',
- 'editor29', 'editor29-first', 'editor29-last', 'editor29link', 'editor29-link',
- 'editor29mask', 'editor29-mask', 'editor30', 'editor30-first', 'editor30-last',
- 'editor30link', 'editor30-link', 'editor30mask', 'editor30-mask', 'editor31',
- 'editor31-first', 'editor31-last', 'editor31link', 'editor31-link', 'editor31mask',
- 'editor31-mask', 'editor32', 'editor32-first', 'editor32-last', 'editor32link',
- 'editor32-link', 'editor32mask', 'editor32-mask', 'editor33', 'editor33-first',
- 'editor33-last', 'editor33link', 'editor33-link', 'editor33mask', 'editor33-mask',
- 'editor34', 'editor34-first', 'editor34-last', 'editor34link', 'editor34-link',
- 'editor34mask', 'editor34-mask', 'editor35', 'editor35-first', 'editor35-last',
- 'editor35link', 'editor35-link', 'editor35mask', 'editor35-mask', 'editor36',
- 'editor36-first', 'editor36-last', 'editor36link', 'editor36-link', 'editor36mask',
- 'editor36-mask', 'editor37', 'editor37-first', 'editor37-last', 'editor37link',
- 'editor37-link', 'editor37mask', 'editor37-mask', 'editor38', 'editor38-first',
- 'editor38-last', 'editor38link', 'editor38-link', 'editor38mask', 'editor38-mask',
- 'editor39', 'editor39-first', 'editor39-last', 'editor39link', 'editor39-link',
- 'editor39mask', 'editor39-mask', 'editor40', 'editor40-first', 'editor40-last',
- 'editor40link', 'editor40-link', 'editor40mask', 'editor40-mask', 'editor41',
- 'editor41-first', 'editor41-last', 'editor41link', 'editor41-link', 'editor41mask',
- 'editor41-mask', 'editor42', 'editor42-first', 'editor42-last', 'editor42link',
- 'editor42-link', 'editor42mask', 'editor42-mask', 'editor43', 'editor43-first',
- 'editor43-last', 'editor43link', 'editor43-link', 'editor43mask', 'editor43-mask',
- 'editor44', 'editor44-first', 'editor44-last', 'editor44link', 'editor44-link',
- 'editor44mask', 'editor44-mask', 'editor45', 'editor45-first', 'editor45-last',
- 'editor45link', 'editor45-link', 'editor45mask', 'editor45-mask', 'editor-first1',
- 'editor-first2', 'editor-first3', 'editor-first4', 'editor-first5', 'editor-first6',
- 'editor-first7', 'editor-first8', 'editor-first9', 'editor-first10', 'editor-first11',
- 'editor-first12', 'editor-first13', 'editor-first14', 'editor-first15', 'editor-first16',
- 'editor-first17', 'editor-first18', 'editor-first19', 'editor-first20', 'editor-first21',
- 'editor-first22', 'editor-first23', 'editor-first24', 'editor-first25', 'editor-first26',
- 'editor-first27', 'editor-first28', 'editor-first29', 'editor-first30', 'editor-first31',
- 'editor-first32', 'editor-first33', 'editor-first34', 'editor-first35', 'editor-first36',
- 'editor-first37', 'editor-first38', 'editor-first39', 'editor-first40', 'editor-first41',
- 'editor-first42', 'editor-first43', 'editor-first44', 'editor-first45', 'editor-last1',
- 'editor-last2', 'editor-last3', 'editor-last4', 'editor-last5', 'editor-last6',
- 'editor-last7', 'editor-last8', 'editor-last9', 'editor-last10', 'editor-last11',
- 'editor-last12', 'editor-last13', 'editor-last14', 'editor-last15', 'editor-last16',
- 'editor-last17', 'editor-last18', 'editor-last19', 'editor-last20', 'editor-last21',
- 'editor-last22', 'editor-last23', 'editor-last24', 'editor-last25', 'editor-last26',
- 'editor-last27', 'editor-last28', 'editor-last29', 'editor-last30', 'editor-last31',
- 'editor-last32', 'editor-last33', 'editor-last34', 'editor-last35', 'editor-last36',
- 'editor-last37', 'editor-last38', 'editor-last39', 'editor-last40', 'editor-last41',
- 'editor-last42', 'editor-last43', 'editor-last44', 'editor-last45', 'editorlink1',
- 'editorlink2', 'editorlink3', 'editorlink4', 'editorlink5', 'editorlink6',
- 'editorlink7', 'editorlink8', 'editorlink9', 'editorlink10', 'editorlink11',
- 'editorlink12', 'editorlink13', 'editorlink14', 'editorlink15', 'editorlink16',
- 'editorlink17', 'editorlink18', 'editorlink19', 'editorlink20', 'editorlink21',
- 'editorlink22', 'editorlink23', 'editorlink24', 'editorlink25', 'editorlink26',
- 'editorlink27', 'editorlink28', 'editorlink29', 'editorlink30', 'editorlink31',
- 'editorlink32', 'editorlink33', 'editorlink34', 'editorlink35', 'editorlink36',
- 'editorlink37', 'editorlink38', 'editorlink39', 'editorlink40', 'editorlink41',
- 'editorlink42', 'editorlink43', 'editorlink44', 'editorlink45', 'editor-link1',
- 'editor-link2', 'editor-link3', 'editor-link4', 'editor-link5', 'editor-link6',
- 'editor-link7', 'editor-link8', 'editor-link9', 'editor-link10', 'editor-link11',
- 'editor-link12', 'editor-link13', 'editor-link14', 'editor-link15', 'editor-link16',
- 'editor-link17', 'editor-link18', 'editor-link19', 'editor-link20', 'editor-link21',
- 'editor-link22', 'editor-link23', 'editor-link24', 'editor-link25', 'editor-link26',
- 'editor-link27', 'editor-link28', 'editor-link29', 'editor-link30', 'editor-link31',
- 'editor-link32', 'editor-link33', 'editor-link34', 'editor-link35', 'editor-link36',
- 'editor-link37', 'editor-link38', 'editor-link39', 'editor-link40', 'editor-link41',
- 'editor-link42', 'editor-link43', 'editor-link44', 'editor-link45', 'editormask1',
- 'editormask2', 'editormask3', 'editormask4', 'editormask5', 'editormask6',
- 'editormask7', 'editormask8', 'editormask9', 'editormask10', 'editormask11',
- 'editormask12', 'editormask13', 'editormask14', 'editormask15', 'editormask16',
- 'editormask17', 'editormask18', 'editormask19', 'editormask20', 'editormask21',
- 'editormask22', 'editormask23', 'editormask24', 'editormask25', 'editormask26',
- 'editormask27', 'editormask28', 'editormask29', 'editormask30', 'editormask31',
- 'editormask32', 'editormask33', 'editormask34', 'editormask35', 'editormask36',
- 'editormask37', 'editormask38', 'editormask39', 'editormask40', 'editormask41',
- 'editormask42', 'editormask43', 'editormask44', 'editormask45', 'editor-mask1',
- 'editor-mask2', 'editor-mask3', 'editor-mask4', 'editor-mask5', 'editor-mask6',
- 'editor-mask7', 'editor-mask8', 'editor-mask9', 'editor-mask10', 'editor-mask11',
- 'editor-mask12', 'editor-mask13', 'editor-mask14', 'editor-mask15', 'editor-mask16',
- 'editor-mask17', 'editor-mask18', 'editor-mask19', 'editor-mask20', 'editor-mask21',
- 'editor-mask22', 'editor-mask23', 'editor-mask24', 'editor-mask25', 'editor-mask26',
- 'editor-mask27', 'editor-mask28', 'editor-mask29', 'editor-mask30', 'editor-mask31',
- 'editor-mask32', 'editor-mask33', 'editor-mask34', 'editor-mask35', 'editor-mask36',
- 'editor-mask37', 'editor-mask38', 'editor-mask39', 'editor-mask40', 'editor-mask41',
- 'editor-mask42', 'editor-mask43', 'editor-mask44', 'editor-mask45', 'editors1',
- 'editors2', 'editors3', 'editors4', 'editors5', 'editors6',
- 'editors7', 'editors8', 'editors9', 'editors10', 'editors11',
- 'editors12', 'editors13', 'editors14', 'editors15', 'editors16',
- 'editors17', 'editors18', 'editors19', 'editors20', 'editors21',
- 'editors22', 'editors23', 'editors24', 'editors25', 'editors26',
- 'editors27', 'editors28', 'editors29', 'editors30', 'editors31',
- 'editors32', 'editors33', 'editors34', 'editors35', 'editors36',
- 'editors37', 'editors38', 'editors39', 'editors40', 'editors41',
- 'editors42', 'editors43', 'editors44', 'editors45', 'first1',
- 'first2', 'first3', 'first4', 'first5', 'first6',
- 'first7', 'first8', 'first9', 'first10', 'first11',
- 'first12', 'first13', 'first14', 'first15', 'first16',
- 'first17', 'first18', 'first19', 'first20', 'first21',
- 'first22', 'first23', 'first24', 'first25', 'first26',
- 'first27', 'first28', 'first29', 'first30', 'first31',
- 'first32', 'first33', 'first34', 'first35', 'first36',
- 'first37', 'first38', 'first39', 'first40', 'first41',
- 'first42', 'first43', 'first44', 'first45', 'first46',
- 'first47', 'first48', 'first49', 'first50', 'first51',
- 'first52', 'first53', 'first54', 'first55', 'first56',
- 'first57', 'first58', 'first59', 'first60', 'first61',
- 'first62', 'first63', 'first64', 'first65', 'first66',
- 'first67', 'first68', 'first69', 'first70', 'first71',
- 'first72', 'first73', 'first74', 'first75', 'first76',
- 'first77', 'first78', 'first79', 'first80', 'first81',
- 'first82', 'first83', 'first84', 'first85', 'first86',
- 'first87', 'first88', 'first89', 'first90', 'first91',
- 'first92', 'first93', 'first94', 'first95', 'first96',
- 'first97', 'first98', 'first99', 'given1', 'given2',
- 'given3', 'given4', 'given5', 'given6', 'given7',
- 'given8', 'given9', 'given10', 'given11', 'given12',
- 'given13', 'given14', 'given15', 'given16', 'given17',
- 'given18', 'given19', 'given20', 'last1', 'last2',
- 'last3', 'last4', 'last5', 'last6', 'last7',
- 'last8', 'last9', 'last10', 'last11', 'last12',
- 'last13', 'last14', 'last15', 'last16', 'last17',
- 'last18', 'last19', 'last20', 'last21', 'last22',
- 'last23', 'last24', 'last25', 'last26', 'last27',
- 'last28', 'last29', 'last30', 'last31', 'last32',
- 'last33', 'last34', 'last35', 'last36', 'last37',
- 'last38', 'last39', 'last40', 'last41', 'last42',
- 'last43', 'last44', 'last45', 'last46', 'last47',
- 'last48', 'last49', 'last50', 'last51', 'last52',
- 'last53', 'last54', 'last55', 'last56', 'last57',
- 'last58', 'last59', 'last60', 'last61', 'last62',
- 'last63', 'last64', 'last65', 'last66', 'last67',
- 'last68', 'last69', 'last70', 'last71', 'last72',
- 'last73', 'last74', 'last75', 'last76', 'last77',
- 'last78', 'last79', 'last80', 'last81', 'last82',
- 'last83', 'last84', 'last85', 'last86', 'last87',
- 'last88', 'last89', 'last90', 'last91', 'last92',
- 'last93', 'last94', 'last95', 'last96', 'last97',
- 'last98', 'last99', 'subject1', 'subject2', 'subject3',
- 'subject4', 'subject5', 'subject6', 'subject7', 'subject8',
- 'subject9', 'subject10', 'subject11', 'subject12', 'subject13',
- 'subject14', 'subject15', 'subject16', 'subject17', 'subject18',
- 'subject19', 'subject20', 'subjectlink1', 'subjectlink2', 'subjectlink3',
- 'subjectlink4', 'subjectlink5', 'subjectlink6', 'subjectlink7', 'subjectlink8',
- 'subjectlink9', 'subjectlink10', 'subjectlink11', 'subjectlink12', 'subjectlink13',
- 'subjectlink14', 'subjectlink15', 'subjectlink16', 'subjectlink17', 'subjectlink18',
- 'subjectlink19', 'subjectlink20', 'surname1', 'surname2', 'surname3',
- 'surname4', 'surname5', 'surname6', 'surname7', 'surname8',
- 'surname9', 'surname10', 'surname11', 'surname12', 'surname13',
- 'surname14', 'surname15', 'surname16', 'surname17', 'surname18',
- 'surname19', 'surname20', 'inventor-surname', 'inventor-surname1', 'inventor-surname2',
- 'inventor-surname3', 'inventor-surname4', 'inventor-surname5', 'inventor-surname6', 'inventor-surname7',
- 'inventor-surname8', 'inventor-surname9', 'inventor-surname10', 'inventor-surname11', 'inventor-surname12',
- 'inventor-surname13', 'inventor-surname14', 'inventor-surname15', 'inventor-surname16', 'inventor-surname17',
- 'inventor-surname18', 'inventor-surname19', 'inventor-surname20', 'inventor-last', 'inventor-last1',
- 'inventor-last2', 'inventor-last3', 'inventor-last4', 'inventor-last5', 'inventor-last6',
- 'inventor-last7', 'inventor-last8', 'inventor-last9', 'inventor-last10', 'inventor-last11',
- 'inventor-last12', 'inventor-last13', 'inventor-last14', 'inventor-last15', 'inventor-last16',
- 'inventor-last17', 'inventor-last18', 'inventor-last19', 'inventor-last20', 'inventor-first',
- 'inventor-first1', 'inventor-first2', 'inventor-first3', 'inventor-first4', 'inventor-first5',
- 'inventor-first6', 'inventor-first7', 'inventor-first8', 'inventor-first9', 'inventor-first10',
- 'inventor-first11', 'inventor-first12', 'inventor-first13', 'inventor-first14', 'inventor-first15',
- 'inventor-first16', 'inventor-first17', 'inventor-first18', 'inventor-first19', 'inventor-first20',
- 'inventor-given', 'inventor-given1', 'inventor-given2', 'inventor-given3', 'inventor-given4',
- 'inventor-given5', 'inventor-given6', 'inventor-given7', 'inventor-given8', 'inventor-given9',
- 'inventor-given10', 'inventor-given11', 'inventor-given12', 'inventor-given13', 'inventor-given14',
- 'inventor-given15', 'inventor-given16', 'inventor-given17', 'inventor-given18', 'inventor-given19',
- 'inventor-given20', 'inventorlink', 'inventorlink1', 'inventorlink2', 'inventorlink3',
- 'inventorlink4', 'inventorlink5', 'inventorlink6', 'inventorlink7', 'inventorlink8',
- 'inventorlink9', 'inventorlink10', 'inventorlink11', 'inventorlink12', 'inventorlink13',
- 'inventorlink14', 'inventorlink15', 'inventorlink16', 'inventorlink17', 'inventorlink18',
- 'inventorlink19', 'inventorlink20', 'patent-number', 'publication-number', 'description',
- 'pubdate', 'issue-date', 'gdate', 'fdate', 'pridate',
- 'assign1', 'assign2', 'inventor-last', 'inventor-last1', 'inventor-last2',
- 'inventor-last3', 'inventor-last4', 'inventor-last5', 'inventor-last6', 'inventor-last7',
- 'inventor-last8', 'inventor-last9', 'inventor-last10', 'inventor-last11', 'inventor-last12',
- 'inventor-last13', 'inventor-last14', 'inventor-last15', 'inventor-last16', 'inventor-last17',
- 'inventor-last18', 'inventor-last19', 'inventor-last20', 'inventor-last21', 'inventor-last22',
- 'inventor-last23', 'inventor-last24', 'inventor-last25', 'inventor-last26', 'inventor-last27',
- 'inventor-last28', 'inventor-last29', 'inventor-last30', 'inventor-first', 'inventor-first1',
- 'inventor-first2', 'inventor-first3', 'inventor-first4', 'inventor-first5', 'inventor-first6',
- 'inventor-first7', 'inventor-first8', 'inventor-first9', 'inventor-first10', 'inventor-first11',
- 'inventor-first12', 'inventor-first13', 'inventor-first14', 'inventor-first15', 'inventor-first16',
- 'inventor-first17', 'inventor-first18', 'inventor-first19', 'inventor-first20', 'inventor-first21',
- 'inventor-first22', 'inventor-first23', 'inventor-first24', 'inventor-first25', 'inventor-first26',
- 'inventor-first27', 'inventor-first28', 'inventor-first29', 'inventor-first30', 'inventor1-last',
- 'inventor2-last', 'inventor3-last', 'inventor4-last', 'inventor5-last', 'inventor6-last',
- 'inventor7-last', 'inventor8-last', 'inventor9-last', 'inventor10-last', 'inventor11-last',
- 'inventor12-last', 'inventor13-last', 'inventor14-last', 'inventor15-last', 'inventor16-last',
- 'inventor17-last', 'inventor18-last', 'inventor19-last', 'inventor20-last', 'inventor21-last',
- 'inventor22-last', 'inventor23-last', 'inventor24-last', 'inventor25-last', 'inventor26-last',
- 'inventor27-last', 'inventor28-last', 'inventor29-last', 'inventor30-last', 'inventor1-first',
- 'inventor2-first', 'inventor3-first', 'inventor4-first', 'inventor5-first', 'inventor6-first',
- 'inventor7-first', 'inventor8-first', 'inventor9-first', 'inventor10-first', 'inventor11-first',
- 'inventor12-first', 'inventor13-first', 'inventor14-first', 'inventor15-first', 'inventor16-first',
- 'inventor17-first', 'inventor18-first', 'inventor19-first', 'inventor20-first', 'inventor21-first',
- 'inventor22-first', 'inventor23-first', 'inventor24-first', 'inventor25-first', 'inventor26-first',
- 'inventor27-first', 'inventor28-first', 'inventor29-first', 'inventor30-first');
-
- const MML_TAGS = array (
-  '<mml:msqrt>'  => "&#8730;(",
-  '<mml:mi>'     => "''",
-  '</mml:mi>'    => "''",
-  '</mml:msqrt>' => ')',
-  '<mml:mo>'     => '',
-  '</mml:mo>'    => '',
-  '<mml:mn>'     => '',
-  '</mml:mn>'    => '',
-  '<mml:mtext>'  => '',
-  '</mml:mtext>' => '',
-  );
-  
  
  ?>
